# src/centralserver/admin-service/src/main/resources/openapi-definition.yaml
openapi: 3.0.0
info:
  description: X-Road Central Server Admin API. Note that the error metadata responses described in some endpoints are subjects to change and may be updated in upcoming versions.
  version: "1.0.0"
  title: X-Road Central Server Admin API
  contact:
    name: Nordic Institute for Interoperability Solutions (NIIS)
    url: https://github.com/nordic-institute/X-Road-development/#enhancement-requests-and-error-reports
    email: info@niis.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: basepath for API v1
paths:
...
  /management-requests:
    get:
      tags:
        - management-requests
      summary: List management requests
      operationId: getManagementRequests
      description: List management requests
      responses:
        '200':
          description: array of management requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementRequestPage'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    post:
      tags:
        - management-requests
      summary: Add a management request
      operationId: addManagementRequest
      description: Add management request
      requestBody:
        description: management request to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementRequest'
      responses:
        '200':
          description: the added management request
          content:
            application/json:
              schema:
                type: integer
                description: management request id
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
  /management-requests/{id}:
    delete:
      tags:
        - management-requests
      summary: Revoke/decline a management request
      operationId: revokeManagementRequest
      description: Administrator revokes a management request
      parameters:
        - in: path
          name: id
          description: request id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Revoke succeeded
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
components:
  schemas:
    Page:
      type: object
      properties:
        number:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
...
    XRoadId:
      type: object
      x-abstract: true
      description: X-Road identifier
      discriminator:
        propertyName: type
        mapping:
          MEMBER: "MemberId"
          SECURITYSERVER: "SecurityServerId"
      properties:
        instanceId:
          type: string
        type:
          type: string
          format: enum
          enum:
            - MEMBER
            - SECURITYSERVER
...
    SecurityServerId:
      allOf:
        - $ref: '#/components/schemas/XRoadId'
      type: object
      properties:
        memberClass:
          type: string
        memberCode:
          type: string
        serverCode:
          type: string
    ManagementRequest:
      type: object
      description: Management request
      discriminator:
        propertyName: requestType
      properties:
        securityserverId:
          $ref: '#/components/schemas/SecurityServerId'
        source:
          type: string
          description: request source
          enum:
            - CENTER
            - SECURITY_SERVER
      required:
        - requestType
        - source
        - securityserverId
    ManagementRequestPage:
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ManagementRequest'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: X-Road-ApiKey token=<api key>
      # https://github.com/nordic-institute/X-Road-REST-UI/blob/XRDDEV-237/doc/Manuals/ug-ss_x-road_6_security_server_user_guide.md#1911-creating-new-api-keys
...
security:
  - ApiKeyAuth: [ ]